Create EventTracker class
Create EventTracker class, to keep track of match events and expose a
function `findNewEvents` which receives an array of the latest matches
data and compares it the the stored data and returns the new events that
should be notified to the user.
Add tests for EventTracker.
Ignore dist directory when running jest tests.

Rename EventStore to EventTracker

minor changes after review.